+++ projects/rocprofiler-compute/src/utils/parser.py
@@ -8,18 +8,18 @@ from typing import Any, Optional
 
 import pandas as pd
 
-from utils import schema
-from utils.logger import console_error, console_warning, demarcate
-from utils.metrics.evaluation_pipeline import eval_metric
-from utils.metrics.expression import gen_counter_list
-from utils.pattern_matching import fnmatch_glob_matches
-from utils.pc_sampling_analysis import (
+from pc_sampling.pc_sampling_analysis import (
     SOURCE_LINE_MISSING,
     aggregate_pc_sample_records,
     detect_pc_sampling_method,
     enrich_with_metadata,
     load_pc_sample_records,
 )
+from utils import schema
+from utils.logger import console_error, console_warning, demarcate
+from utils.metrics.evaluation_pipeline import eval_metric
+from utils.metrics.expression import gen_counter_list
+from utils.pattern_matching import fnmatch_glob_matches
 from utils.specs import MachineSpecs
 from utils.utils_common import (
     METRIC_ID_RE,
